--- app/.env_template
@@ -1,6 +1,0 @@
-jwt_secret=
-jwt_algorithm=
-
-db_psswd=
-db_host=
-db_port=
