--- src/auth/jwt_bearer.py
@@ -1,25 +1,0 @@
-from fastapi import Request, HTTPException
-from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
-from auth.jwt_handler import decodeJWT
-
-class jwtBearer(HTTPBearer):
-    def __init__(self, auto_Error: bool = True):
-        super(jwtBearer,self).__init__(auto_error=auto_Error)
-
-    async def __call__(self, request: Request):
-        credentials: HTTPAuthorizationCredentials = await super(jwtBearer,self).__call__(request)
-        if credentials:
-            if not credentials.scheme == "Bearer":
-                raise HTTPException(status_code=403, detail="Invalid or Expired Token!")
-            if not self.verify_jwt(credentials.credentials):
-                raise HTTPException(status_code=403, detail="Invalid or Expired Token!")
-            return credentials.credentials
-        else:
-            raise HTTPException(status_code=403, detail="Invalid or Expired Token!")
-
-    def verify_jwt(self, jwt_token: str) -> bool:
-        isTokenValid: bool = False
-        payload = decodeJWT(jwt_token)
-        if payload:
-            isTokenValid = True
-        return isTokenValid
