+++ src/init.py
@@ -15,11 +15,12 @@ def initDb():
     cursor.execute(f"""
     CREATE TABLE IF NOT EXISTS users (
         uid text PRIMARY KEY,
-        psswd_hash text,
+        psswd_hash text NOT NULL,
+        can_make_transactions bool NOT NULL,
         is_admin boolean NOT NULL
     );
-    INSERT INTO users (uid, psswd_hash, is_admin)
-    SELECT 'admin','{admin_pwd}',true
+    INSERT INTO users (uid, psswd_hash,can_make_transactions,is_admin)
+    SELECT 'admin','{admin_pwd}',true,true
     WHERE NOT EXISTS (SELECT * FROM users);
     CREATE TABLE IF NOT EXISTS budgets (
         uid text PRIMARY KEY,
